Улучшение стилистики кода и отладка работы проекта

+++ bookkeeper/presenter/category_presenter.py
@@ -10,7 +10,7 @@ class CategoryPresenter:
         self.view = view
         self.view.category_add_button_clicked(self.handle_category_add_button_clicked)
         self.view.category_edit_button_clicked(self.handle_category_edit_button_clicked)
-        self.view.category_delete_button_clicked(self.handle_category_delete_button_clicked)
+        self.view.category_delete_button_clicked(self.handle_category_del_button_clicked)
 
     def show(self) -> None:
         self.view.show()
@@ -26,17 +26,18 @@ class CategoryPresenter:
         self.show()
 
     def handle_category_edit_button_clicked(self) -> None:
-        selected = self.view.get_selected_cat()
-        new_category_name = self.view.get_category_name()
-        cat = Category(name=str(new_category_name), parent=self.cat_repo.get(selected).parent, pk=selected)
+        select = self.view.get_selected_cat()
+        new_name = self.view.get_category_name()
+        cat = Category(name=str(new_name), parent=self.cat_repo.get(select).parent, pk=select)
         self.cat_repo.update(cat)
         self.show()
 
-    def handle_category_delete_button_clicked(self) -> None:
+    def handle_category_del_button_clicked(self) -> None:
         selected = self.view.get_selected_cat()
         if selected:
-            for e in self.exp_repo.get_all():
-                if e.category == selected:
-                    self.exp_repo.delete(e.pk)
+            if self.exp_repo.get_all():
+                for e in self.exp_repo.get_all():
+                    if e.category == selected:
+                        self.exp_repo.delete(e.pk)
             self.cat_repo.delete(selected)
         self.show()
